Project 2 - Analyzing Data with Python
Finished the project to examine a table using panda, while displaying it's information into a graphic with plotly. As well as examining the possible causes for the number of cancellations and the result:
Of when 1 of each of theses causes are solved.
Of when a combinations of 3 of theses causes are solved.
Finding out the best 3 causes to be tackled to reduce the cancellation rate from 55.6% to 14.9%, almost a quarter of the original value.